:truck: Move ProtectedRoute
